feat and fix: update AdminService to use DTOs for creating trails and scenarios, and add error handling for deletion

+++ src/main/java/com/projeto/tcc/letramento/service/AdminService.java
@@ -1,11 +1,15 @@
 package com.projeto.tcc.letramento.service;
 
+import com.projeto.tcc.letramento.dto.ScenarioRequestDTO;
+import com.projeto.tcc.letramento.dto.TrailRequestDTO;
 import com.projeto.tcc.letramento.model.Scenario;
 import com.projeto.tcc.letramento.model.Trail;
 import com.projeto.tcc.letramento.repository.ScenarioRepository;
 import com.projeto.tcc.letramento.repository.TrailRepository;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.jspecify.annotations.NonNull;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -16,17 +20,32 @@ public class AdminService {
     private final ScenarioRepository scenarioRepository;
 
     @Transactional
-    public Trail createTrail(Trail trail) {
+    public Trail createTrail(@NonNull TrailRequestDTO data) {
+        Trail trail = new Trail();
+        trail.setTitle(data.title());
+        trail.setDescription(data.description());
         return trailRepository.save(trail);
     }
 
     @Transactional
-    public Scenario createScenario(Scenario scenario) {
+    public Scenario createScenario(@NonNull ScenarioRequestDTO data) {
+        Trail trail = trailRepository.findById(data.trailId())
+                .orElseThrow(() -> new EntityNotFoundException("Trilha não encontrada"));
+        Scenario scenario = new Scenario();
+        scenario.setTitleScenarios(data.titleScenario());
+        scenario.setXrayData(data.xrayData());
+        scenario.setQuiz(data.quiz());
+        scenario.setPillar(data.pillar());
+        scenario.setTrail(trail);
+
         return scenarioRepository.save(scenario);
     }
 
     @Transactional
     public void deleteTrail(Long trailId) {
+        if (!trailRepository.existsById(trailId)) {
+            throw new EntityNotFoundException("Trilha não encontrada");
+        }
         trailRepository.deleteById(trailId);
     }
 }
